consolidated model provider util into the arthur wrapper file
